+++ webviz_config/testing/__init__.py
@@ -1,0 +1,1 @@
+from ._composite import WebvizComposite